+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Settings tab.
 - Repository tab.
+- Add, delete and modify task.
 
 ## [0.8.0] - 2022-12-24